eslint

+++ src/Osvita/TargetOsvita.js
@@ -8,14 +8,13 @@ const Target = require('../Target')
 const makeFile = function (target) {
     return new Promise((resolve, reject) => {
         try {
-            if (!target.append)
-                removeFile(target.fullFileName)
+            if (!target.append) { removeFile(target.fullFileName) }
             fs.exists(target.sourceFullFileName, (exists) => {
                 if (exists) {
                     let buffer = target.append ? '' : target.entity.getHeader()
                     let id = 1
                     fs.createReadStream(target.sourceFullFileName)
-                        .pipe(new YADBF({ 
+                        .pipe(new YADBF({
                             // encoding: 'cp1251'
                             encoding: 'cp866'
                         }))
@@ -37,7 +36,7 @@ const makeFile = function (target) {
                             resolve(target)
                         })
                         .on('error', err => {
-                            console.error(`an error was thrown: ${err}`);
+                            console.error(`an error was thrown: ${err}`)
                             target.state = Target.FILE_ERROR
                             target.err = err.message
                             resolve(target)
